Fixed bug 3579434.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@4780 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/platforms/STM32/USARTv2/uart_lld.h
@@ -340,15 +340,15 @@ typedef struct {
   /**
    * @brief Initialization value for the CR1 register.
    */
-  uint16_t                  cr1;
+  uint32_t                  cr1;
   /**
    * @brief Initialization value for the CR2 register.
    */
-  uint16_t                  cr2;
+  uint32_t                  cr2;
   /**
    * @brief Initialization value for the CR3 register.
    */
-  uint16_t                  cr3;
+  uint32_t                  cr3;
 } UARTConfig;
 
 /**
